Method_2/requirements.txt
langchain
langchain-google-genai
python-dotenv
tk 